added load data for authors, refactor toggle authors, memorize author card for less renders
